refactor: reorganize project structure into separate packages for controllers, services, repositories, dtos, and mappers
