+++ src/main/java/com/gutkoski/trustscore/dto/ProductDTO.java
@@ -1,0 +1,4 @@
+package com.gutkoski.trustscore.dto;
+
+public class ProductDTO {
+}
