+++ src/main/java/com/gutkoski/trustscore/dto/UserResponseDTO.java
@@ -1,4 +1,4 @@
-package com.gutkoski.trustscore.entities.user.dto;
+package com.gutkoski.trustscore.dto;
 
 public record UserResponseDTO(Long id,
                               String name,
