--- src/main/java/com/gutkoski/trustscore/entities/product/ProductDTO.java
@@ -1,4 +1,0 @@
-package com.gutkoski.trustscore.entities.product;
-
-public class ProductDTO {
-}
