+++ src/main/java/com/gutkoski/trustscore/repository/ReviewRepository.java
@@ -1,5 +1,6 @@
-package com.gutkoski.trustscore.entities.review;
+package com.gutkoski.trustscore.repository;
 
+import com.gutkoski.trustscore.entity.Review;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
