+++ src/main/java/com/gutkoski/trustscore/repository/UserRepository.java
@@ -1,5 +1,6 @@
-package com.gutkoski.trustscore.entities.user;
+package com.gutkoski.trustscore.repository;
 
+import com.gutkoski.trustscore.entity.User;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
